+++ common/src/python/flywheel_adaptor/flywheel_proxy.py
@@ -5,7 +5,7 @@ import json
 import logging
 from codecs import StreamReader
 from json.decoder import JSONDecodeError
-from typing import Any, Dict, Iterable, List, Literal, Mapping, Optional
+from typing import Any, Dict, Iterable, List, Literal, Mapping, Optional, Sequence
 
 import flywheel
 from flywheel import (
@@ -327,6 +327,26 @@ class FlywheelProxy:
         """
         role_map = self.get_roles()
         return role_map.get(label)
+
+    def get_user_roles(self, role_names: Sequence[str]) -> List[RoleOutput]:
+        """Get the named roles .
+
+        Returns all roles matching a name in the list.
+        Logs a warning if a name is not matched.
+
+        Args:
+          role_names: the role names
+        Returns:
+          the list of roles with the names
+        """
+        role_list = []
+        for name in role_names:
+            role = self.get_role(name)
+            if role:
+                role_list.append(role)
+            else:
+                log.warning("no such role %s", name)
+        return role_list
 
     def get_admin_role(self) -> Optional[RoleOutput]:
         """Gets admin role."""
@@ -741,6 +761,9 @@ class FlywheelProxy:
         """
         return self.__fw.get(container_id)
 
+    def find_projects_with_pattern(self, pattern: str) -> List[flywheel.Project]:
+        return self.__fw.projects.find(f"label=~{pattern}")
+
 
 def get_name(container) -> str:
     """Returns the name for the container.
@@ -916,8 +939,10 @@ class GroupAdaptor:
 
         self._fw.add_group_role(group=self._group, role=new_role)
 
-    def add_roles(self, roles: List[GroupRole]) -> None:
+    def add_roles(self, roles: Sequence[GroupRole]) -> None:
         """Adds the roles in the list to the group.
+
+        Allows roles to be assigned to users for projects w/in the group.
 
         Args:
           roles: the list of roles
@@ -986,7 +1011,7 @@ class ProjectAdaptor:
     """Defines an adaptor for a flywheel project."""
 
     def __init__(self, *, project: flywheel.Project, proxy: FlywheelProxy) -> None:
-        self._project = project
+        self._project = project.reload()
         self._fw = proxy
 
     @classmethod
@@ -1042,6 +1067,29 @@ class ProjectAdaptor:
     def group(self) -> str:
         """Returns the group label of the enclosed project."""
         return self._project.group
+
+    def get_pipeline_adcid(self) -> int:
+        """Returns the pipeline ADCID for this project.
+
+        If not in the project metadata, the ADCID from the metadata is returned.
+        If neither is present raises an exception.
+
+        Returns:
+          the pipeline ADCID
+        Raises:
+          ProjectError if neither "pipeline_adcid" or "adcid" are set in
+          project metadata.
+        """
+
+        pipeline_adcid = self._project.info.get("pipeline_adcid")
+        if pipeline_adcid is not None:
+            return pipeline_adcid
+
+        adcid = self._project.info.get("adcid")
+        if adcid is not None:
+            return adcid
+
+        raise ProjectError(f"Project {self.group}/{self.label} has no ADCID")
 
     def add_tag(self, tag: str) -> None:
         """Add tag to the enclosed project.