+++ docs/legacy_identifier_transfer/CHANGELOG.md
@@ -2,7 +2,9 @@
 
 All notable changes to this gear are documented in this file.
 
-## 1.2.0
+## 2.0.0
+* Rebuilt for ssm-parameter-store update
+* Changes to use the ADCID from the parent project custom info for the pipeline ADCID. Fails if the ADCID is not set.
 * Updates to use new identifier lambda functions with restructured identifiers database
   
 ## 1.1.6